+++ lib/components/atoms/list/list-item/index.ts
@@ -1,1 +1,2 @@
-export { default } from './list-item';
+export { default } from "./list-item";
+export type { ListItemProps } from "./list-item";