+++ lib/components/atoms/list/list-item/list-item.tsx
@@ -1,6 +1,6 @@
-import React, { Component } from 'react';
-import styles from '../list.module.css';
-import clsx from 'clsx';
+import React, { Component } from "react";
+import styles from "../list.module.css";
+import clsx from "clsx";
 
 /*
 is my child an anchor ?!
@@ -10,8 +10,24 @@ is my child an anchor ?!
   </li>
 */
 
-const ListItem = ({ children, customClass = '', customStyle = {}, ...rest }) => (
-  <li {...rest} className={clsx(styles.listItem, customClass)} style={customStyle}>
+export interface ListItemProps {
+  children: React.ReactNode;
+  customClass?: string;
+  customStyle?: React.CSSProperties;
+  role?: string;
+}
+
+const ListItem = ({
+  children,
+  customClass = "",
+  customStyle = {},
+  ...rest
+}: ListItemProps) => (
+  <li
+    {...rest}
+    className={clsx(styles.listItem, customClass)}
+    style={customStyle}
+  >
     {children}
   </li>
 );