+++ lib/components/organisms/branches/__tests__/branch-list.test.tsx
@@ -1,21 +1,28 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import { createAllPortholeTrees, getRSSBranch, convertToPortholeBranches } from '../../../../utils/golds/feed-util';
-import { getAllItemsFromStore } from '../../../../utils/golds/indexdb-util';
-import { differenceBy, shuffle } from '../../../../utils/fp-util';
+import React from "react";
+import { render } from "@testing-library/react";
+import {
+  createAllPortholeTrees,
+  getRSSBranch,
+  convertToPortholeBranches,
+} from "@cross-country/utils/golds/feed-util";
+import { getAllItemsFromStore } from "@cross-country/utils/golds/indexdb-util";
+import { differenceBy, shuffle } from "@cross-country/utils/fp-util";
+import PortholeBranchModel from "@cross-country/models/PortholeBranchModel";
 
-import { BranchType } from '../../../../types/branch.type';
-import { BranchListProps } from '../branch-list.types';
-import { response } from '../__mocks__/branches-mock';
-import BranchList from '../branch-list';
+import { BranchType } from "@cross-country/types/branch.type";
+import { BranchListProps } from "../branch-list";
+import { response } from "../__mocks__/branches-mock";
+import BranchList from "../branch-list";
 
 const allNewBranches = convertToPortholeBranches(response.feed_responses);
 const shuffledBranches = shuffle(allNewBranches);
 
 // ensure all branches are unique
-const uniqueBranches = [...new Set<BranchType>(shuffledBranches)];
+const uniqueBranches = [
+  ...new Set<PortholeBranchModel>(shuffledBranches),
+] as PortholeBranchModel[];
 
-describe('Test Component', () => {
+describe("Test Component", () => {
   let props: BranchListProps;
 
   beforeEach(() => {
@@ -26,10 +33,10 @@ describe('Test Component', () => {
 
   const renderComponent = () => render(<BranchList {...props} />);
 
-  it('should render foo text correctly', () => {
+  it("should render foo text correctly", () => {
     const { getByTestId } = renderComponent();
 
-    const component = getByTestId('branch-list');
+    const component = getByTestId("branch-list");
 
     expect(component).toBeTruthy();
   });