+++ lib/components/organisms/branches/branch-list.types.ts
@@ -1,5 +1,0 @@
-import { BranchType } from '../../../types/branch.type';
-
-export interface BranchListProps {
-  branches: BranchType[];
-}