+++ lib/models/TreeModel.ts
@@ -1,4 +1,6 @@
-export type TreeType = {
+// TreeModel would be the get/set version using immutable.js
+import { Record } from "immutable";
+export type TreeModelType = {
   _id: string;
   xmlUrl: string;
   type: string;
@@ -6,3 +8,14 @@ export type TreeType = {
   category: string;
   origin: string;
 };
+
+const TreeModel = Record<TreeModelType>({
+  _id: "",
+  xmlUrl: "",
+  type: "",
+  title: "",
+  category: "",
+  origin: "",
+});
+
+export default TreeModel;